Moved ActionRouter interface
