+++ src/app.py
@@ -1,6 +1,5 @@
 from common import *
 from routes import *
-from actions import action_router
 
 async def main():
     state_task = None
